+++ src/tabuleiro.cpp
@@ -55,7 +55,7 @@ std::vector<Peca_t> Tabuleiro_t::get_tabuleiro(){
     return _tabuleiro;
 }
 
-bool Tabuleiro_t::procura_peca(int linha, int coluna, char cor) {
+bool Tabuleiro_t::procura_peca(long unsigned int linha, long unsigned int coluna, char cor) {
     for (Peca_t peca : _tabuleiro) {
         if (peca.get_posicao().linha == linha && 
            peca.get_posicao().coluna == coluna &&
